OOP folder added

+++ Fundamentals/DeepDive_Fibonacci/Fibonacci.java
@@ -1,12 +1,12 @@
 class Fibonacci{
 	public static void main(String[] args) {
-		int[] array = new int[100];
+		long[] array = new long[100];
 		populateArray(array);
 		printArray(array);
 	}
 
 	//populating array
-	public static void populateArray(int[] array){
+	public static void populateArray(long[] array){
 		array[0] = 0;
 		array[1] = 1;
 		for(int i =2; i<array.length; i++){
@@ -15,8 +15,8 @@ class Fibonacci{
 	}
 
 	//printing array
-	public static void printArray(int[] array){
-		for(int element: array){
+	public static void printArray(long[] array){
+		for(long element: array){
 			System.out.println(element);
 		}
 	}
